+++ fns-test/userInfo.test.js
@@ -8,22 +8,18 @@ import { strict as assert } from 'assert'
 
 import { test, expect, describe, beforeAll, afterAll, jest } from '@jest/globals'
 
-import { db } from './tools/session.js'
-import { extTimeoutPromise } from "./tools/extPromise"
+import { db } from './tools/firebase.js'
+import { bestBeforePromise } from "./tools/promiseTools.js"
 
 import './matchers/toContainObject'
-
-import { performance } from 'perf_hooks'    // node.js
-
 
 // Clear '/projects/1/userInfo/abc'
 //
 // Note: Since we leave the end result of running tests in the server, this is needed in case (in dev) the test is
 //    run multiple times.
 //
-beforeAll( async () => {
+beforeAll( async () => {    // takes about 456, 419 ms
 
-  const t0 = performance.now();
   try {
     await db.doc("projects/1/userInfo/abc").delete();    // left-overs from earlier tests - maybe
 
@@ -34,20 +30,11 @@ beforeAll( async () => {
     console.error("Initialization failed:", err);
     throw err;    // tbd. how to NOT run tests if 'beforeAll' fails? #jest
   }
-
-  console.debug("Cleared earlier changes (if any). Took:", (performance.now() - t0).toFixed(2) + "ms")
 });
 
 afterAll( async () => {
   await db.app.delete();
-
-  console.debug("Cleanup finished (leaving data changes in the database).")
 });
-
-// tbd. candidate for 'tools/'
-// Wait 'ms', as a Promise
-//
-const waitMs = ms => new Promise(resolve => setTimeout(resolve, ms));
 
 describe("userInfo shadowing", () => {
 
@@ -60,16 +47,14 @@ describe("userInfo shadowing", () => {
       photoURL: "https://upload.wikimedia.org/wikipedia/commons/a/ab/Dalton_Bill-edit.png"
     };
 
-    const exProm = bestBeforePromise({ timeoutMs: 500 });    // { promise: Promise, resolve: () => (), reject: () => () }
-
-    expect.assertions(1);
+    const exProm = bestBeforePromise({ timeoutMs: 200 });    // { promise: Promise, resolve: () => (), reject: () => () }
 
     // Prepare a watch
     //
     const unsub = db.collection("projects/1/userInfo").doc("abc")
       .onSnapshot(dss => {    // enough to see the intention (write to cache)
         const o = dss.data();
-        console.debug("Noticed: ", o);   // DEBUG
+        //console.debug("Noticed: ", o);   // DEBUG
 
         if (o) {   // 'undefined' on first call (initially no doc)
           expect(o).toContainObject(william);
@@ -77,12 +62,12 @@ describe("userInfo shadowing", () => {
         }
       });
 
-    // Write in central -> should cause Cloud Functions to update the above (but in some ms..)
+    // Write in central -> should cause Cloud Functions to update the above (but not immediately)
     //
     await db.collection("userInfo").doc("abc").set(william);
 
     try {
-      await exProm.promise;   // resolved by the database change; or timeout
+      await exProm;   // resolved by the database change or rejected by timeout
     }
     finally {
       unsub();    // tbd. is this the best place? (does it always get called?)
@@ -90,33 +75,30 @@ describe("userInfo shadowing", () => {
   });
 
   test('Central user information is not distributed to a project where the user is not a member', async () => {
-    const exProm = bestBeforePromise({ timeoutMs: 500 });
+    const exProm = bestBeforePromise({ timeoutMs: 200, onTimeout: () => true });
 
     // Prepare a watch (should NOT get called!)
     //
-    const unsub = db.collection("projects/1/userInfo").doc("abc")
+    const unsub = db.collection("projects/1/userInfo").doc("xyz")
       .onSnapshot(dss => {    // enough to see the intention (write to cache)
         const o = dss.data();
-
-        if (o) {
-          exProm.resolve(false);    // shouldn't have reached here
+        if (o) {  // 'undefined'|obj
+          console.debug("Not expecting visitors...", o);   // DEBUG
+          exProm.resolve(false);    // shouldn't have reached here; cuts the wait
         }
       });
 
     // Write in central -> should NOT cause the above to happen
     //
-    await db.collection("userInfo").doc("no-such").set({ name: "blah", photoURL: "https://hoax.png" });
+    await db.collection("userInfo").doc("xyz").set({ name: "blah", photoURL: "https://no-such.png" });
 
-    try {
-      const v = await exProm.promise;   // rejected by (unwanted) database change (leads to failure of test) or resolved by timeout
-      expect( v === undefined );    // timed out
-    }
-    finally {
-      unsub();
-    }
+    // Resolves to true|false
+    expect(await exProm).toBe(true);
+    unsub();
   });
 });
 
+// tbd. move to 'tools/'
 /*
 * Delete all (prior) documents from a collection.
 */
@@ -124,7 +106,7 @@ async function wipe(collection) {   // CollectionReference => Promise of ()
 
   const qss = await collection.get();
   const proms = qss.docs.map( qdss => {
-    console.debug("Deleting:", collection.id +"/"+ qdss.id);
+    //console.debug("Deleting:", collection.id +"/"+ qdss.id);
     return qdss.ref.delete()
   } );
   await Promise.all(proms);