+++ fns-test/CHANGELOG.md
@@ -2,9 +2,10 @@
 
 *Separate changelog from the main project. These kind of live side-by-side.*
 
-## latest
+## 5-Aug-20
 
-- UserInfo test sometimes passes [WIP]
+- Using native ES modules
+- userInfo testing works! 😀
 
 ## 22-Jul-20